convert two more static ints used as flags to smallints

+++ libc/inet/netlinkaccess.h
@@ -63,7 +63,7 @@ struct netlink_handle
 
 #if 0 /* unused code */
 #if __ASSUME_NETLINK_SUPPORT == 0
-extern int __no_netlink_support attribute_hidden;
+extern smallint __no_netlink_support attribute_hidden;
 #else
 # define __no_netlink_support 0
 #endif
